+++ tests/integration_tests/clippath.rs
@@ -7,7 +7,7 @@ fn test_clippath_simple() {
 <defs>
  <clipPath id="cp"><rect wh="10"/></clipPath>
 </defs>
-<rect id="r1" wh="100" clip-path="url(#cp)">
+<rect id="r1" wh="100" clip-path="url(#cp)"/>
 <rect id="r2" surround="#r1"/>
 "##;
     let expected = r#"<rect id="r2" x="0" y="0" width="10" height="10" class="d-surround"/>"#;
@@ -18,7 +18,7 @@ fn test_clippath_simple() {
 <defs>
  <clipPath id="cp"><rect xy="5" wh="10"/></clipPath>
 </defs>
-<rect id="r1" wh="100" clip-path="url(#cp)">
+<rect id="r1" wh="100" clip-path="url(#cp)"/>
 <rect id="r2" surround="#r1"/>
 "##;
     let expected = r#"<rect id="r2" x="5" y="5" width="10" height="10" class="d-surround"/>"#;
@@ -33,7 +33,7 @@ fn test_clippath_partial() {
 <defs>
  <clipPath id="cp"><rect xy="95" wh="10"/></clipPath>
 </defs>
-<rect id="r1" wh="100" clip-path="url(#cp)">
+<rect id="r1" wh="100" clip-path="url(#cp)"/>
 <rect id="r2" surround="#r1"/>
 "##;
     let expected = r#"<rect id="r2" x="95" y="95" width="5" height="5" class="d-surround"/>"#;
@@ -50,7 +50,7 @@ fn test_clippath_composite() {
   <rect xy="10" wh="50 10"/>
  </clipPath>
 </defs>
-<rect id="r1" wh="100" clip-path="url(#cp)">
+<rect id="r1" wh="100" clip-path="url(#cp)"/>
 <rect id="r2" surround="#r1"/>
 "##;
     let expected = r#"<rect id="r2" x="0" y="0" width="60" height="50" class="d-surround"/>"#;
@@ -66,7 +66,7 @@ fn test_clippath_previous() {
   <rect wh="50"/>
  </clipPath>
 </defs>
-<rect id="r1" wh="100" clip-path="url(#cp)">
+<rect id="r1" wh="100" clip-path="url(#cp)"/>
 <rect id="r2" surround="^"/>
 "##;
     let expected = r#"<rect id="r2" x="0" y="0" width="50" height="50" class="d-surround"/>"#;